requirements.txt
requests>=2.31.0
beautifulsoup4
dashscope==1.13.6
python-dotenv
pytz
cloudscraper==1.2.71
producthunt==0.1.4
selenium
webdriver_manager
feedparser>=6.0.11
tenacity>=8.2.2
markdown>=3.4.3
aiohttp>=3.8.4
Pillow>=9.4.0